# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-ready event trigger management system supporting scheduled and API-triggered events with Kafka integration.\n\n## Features\n- **Scheduled Triggers**: One-time and recurring schedules with ISO 8601 and interval-based scheduling\n- **API Triggers**: Webhook-based triggers with JSON schema validation\n- **Event Logs**: Comprehensive logging with retention lifecycle (active → archived → deleted)\n- **Kafka Integration**: Reliable event publishing for external consumers\n\n## Architecture\nThis platform is an event publisher, not an executor. We publish trigger events to Kafka, and external consumers (owned by users) subscribe and execute business logic.",
        "title": "Event Trigger Platform API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Dhi13man/event-trigger-platform",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/events": {
            "get": {
                "description": "Retrieves event logs with filtering and pagination. By default shows only active events (last 2 hours).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "List event logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by trigger ID",
                        "name": "trigger_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "archived"
                        ],
                        "type": "string",
                        "default": "active",
                        "description": "Filter by retention status",
                        "name": "retention_status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "success",
                            "failure"
                        ],
                        "type": "string",
                        "description": "Filter by execution status",
                        "name": "execution_status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "api",
                            "scheduled",
                            "manual-test"
                        ],
                        "type": "string",
                        "description": "Filter by event source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/EventLogListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_api_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{id}": {
            "get": {
                "description": "Retrieves details of a specific event log by ID, including full payload and error message if failed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get event log details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/EventLogResponse"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_api_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "EventLogListResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/EventLogResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/Pagination"
                }
            }
        },
        "EventLogResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-11-05T10:30:00Z"
                },
                "error_message": {
                    "type": "string",
                    "example": "connection timeout"
                },
                "execution_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_models.ExecutionStatus"
                        }
                    ],
                    "example": "success"
                },
                "fired_at": {
                    "type": "string",
                    "example": "2025-11-05T10:30:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                },
                "is_test_run": {
                    "type": "boolean",
                    "example": false
                },
                "payload": {
                    "type": "object"
                },
                "retention_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_models.RetentionStatus"
                        }
                    ],
                    "example": "active"
                },
                "source": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_models.EventSource"
                        }
                    ],
                    "example": "scheduled"
                },
                "trigger_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "trigger_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dhima_event-trigger-platform_internal_models.TriggerType"
                        }
                    ],
                    "example": "scheduled"
                }
            }
        },
...
        "Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                },
                "total_records": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "github_com_dhima_event-trigger-platform_internal_api_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dhima_event-trigger-platform_internal_models.EventSource": {
            "type": "string",
            "enum": [
                "api",
                "scheduled",
                "manual-test"
            ],
            "x-enum-varnames": [
                "EventSourceAPI",
                "EventSourceScheduled",
                "EventSourceManualTest"
            ]
        },
        "github_com_dhima_event-trigger-platform_internal_models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "success",
                "failure"
            ],
            "x-enum-varnames": [
                "ExecutionStatusSuccess",
                "ExecutionStatusFailure"
            ]
        },
        "github_com_dhima_event-trigger-platform_internal_models.RetentionStatus": {
            "type": "string",
            "enum": [
                "active",
                "archived",
                "deleted"
            ],
            "x-enum-varnames": [
                "RetentionStatusActive",
                "RetentionStatusArchived",
                "RetentionStatusDeleted"
            ]
        },
...
        "github_com_dhima_event-trigger-platform_internal_models.TriggerType": {
            "type": "string",
            "enum": [
                "scheduled",
                "api"
            ],
            "x-enum-varnames": [
                "TriggerTypeScheduled",
                "TriggerTypeAPI"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Optional API key authentication for management operations",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}